+++ clang/lib/Driver/ToolChains/Arch/AArch64.cpp
@@ -321,13 +321,8 @@ void aarch64::getAArch64TargetFeatures(const Driver &D,
     }
   }
 
-  if (Arg *A = Args.getLastArg(
-          options::OPT_mstrict_align, options::OPT_mno_strict_align,
-          options::OPT_mno_unaligned_access, options::OPT_munaligned_access)) {
-    if (A->getOption().matches(options::OPT_mstrict_align) ||
-        A->getOption().matches(options::OPT_mno_unaligned_access))
-      Features.push_back("+strict-align");
-  } else if (Triple.isOSOpenBSD())
+  if (Args.hasFlag(options::OPT_mstrict_align, options::OPT_mno_strict_align,
+                   Triple.isOSOpenBSD()))
     Features.push_back("+strict-align");
 
   if (Args.hasArg(options::OPT_ffixed_x1))